requirements.txt
streamlit
paddleocr
opencv-python-headless
Pillow
numpy
pyarrow